+++ packages/plugins/plugin-hono-server/src/index.ts
@@ -2,4 +2,5 @@
 
 export * from './hono-plugin';
 export * from './adapter';
+export * from './pattern-matcher';